+++ src/main/scala/platforms/ulx3s.scala
@@ -8,7 +8,7 @@ import chisel3.util._
 import chisel3.stage.{ChiselGeneratorAnnotation, ChiselStage}
 
 import fpgamacro.gowin.{CLKDIV, TMDS_PLLVR, TLVDS_OBUF}
-import hdmicore.{PatternExample, TMDSDiff, DiffPair}
+import hdmicore.{PatternExample, TMDSDiff, DiffPair, HdmiTx}
 
 class Ulx3s extends RawModule {
 
@@ -58,18 +58,21 @@ class Ulx3s extends RawModule {
       O_led := (counterReg >= (max_count/2).U)
 
       val patternExample = Module(new PatternExample())
-      patternExample.io.serClk := serial_clk
       patternExample.io.I_button := I_button
+
+      val hdmiTx = Module(new HdmiTx())
+      hdmiTx.io.serClk := serial_clk
+      patternExample.io.videoSig <> hdmiTx.io.videoSig
 
       /* LVDS output */
       val buffDiffBlue = Module(new TLVDS_OBUF())
-      buffDiffBlue.io.I := patternExample.io.tmds.data(0)
+      buffDiffBlue.io.I := hdmiTx.io.tmds.data(0)
       val buffDiffGreen = Module(new TLVDS_OBUF())
-      buffDiffGreen.io.I := patternExample.io.tmds.data(1)
+      buffDiffGreen.io.I := hdmiTx.io.tmds.data(1)
       val buffDiffRed = Module(new TLVDS_OBUF())
-      buffDiffRed.io.I := patternExample.io.tmds.data(2)
+      buffDiffRed.io.I := hdmiTx.io.tmds.data(2)
       val buffDiffClk = Module(new TLVDS_OBUF())
-      buffDiffClk.io.I := patternExample.io.tmds.clk
+      buffDiffClk.io.I := hdmiTx.io.tmds.clk
 
       O_tmds.data(0).p := buffDiffBlue.io.O
       O_tmds.data(0).n := buffDiffBlue.io.OB
